Remove `CClient::m_Authed`
As lerato pointed out in b7cbaedea0b72743772c5ce3b43c371e57e4dce1
it is not wise to keep both `m_AuthKey` and `m_Authed` around.
Because they have to be both kept in sync and store the same
information. As you can see in the diff there already seems to
be a case of desync: on timeout only `m_Authed`
was reset but not `m_AuthKey`.

This refactor makes the code simpler, safer and helps with #10681