Easier to read rcon command update code
Before this change the code was non obvious to me.
All the code does is not update the client if nothing changed.

The access levels in general are a bit confusing. Because lower access
levels have more power. So reading 4 comparisons with greater and less
than and these access levels is quite complex.

Also it does nothing if the command was not sent to the client yet.
This check relies on the reader to know that if `m_pRconCmdToSend`
is not `nullptr` the client is currently receiving rcon commands.
And that these commands are sent in alphabetical order.
And that on every send step their access level is checked.
So the `str_comp() >= 0` checks if the changed command was already
sent or not. Because if it was not sent yet we do not need to inform
the client now since it will happen later automatically anyways.

And finally it checks if the commands access level increased or
decreased. But what this really is about is if the client has
access to it or not. Somehow the old check magically worked because
there are not many access levels and because of the huge amount of
prior checks. But it hides the true intention of the code.