Move linked list iterator `NextCommandInfo` to `CConsole`
Needed for #10681

The method to get the first entry is already defined in the `CConsole`
but the next was on the command itself. Which is proper linked list
styled code. And it also worked because the command stores the access
level so while iterating it has all the information it needs.

But when access levels become more complex this will no longer work.
Because it is planned to no longer store that information on the
command. Because there will be multiple different roles that can access
a command and they are not strictly ordered by access level.
So a simple integer comparison will not work.

The alternative would be moving all that context onto the command.
That works by passing one callback and one callback context to every
command on registration. I tried that approach too. It worked.
But I think moving the `NextCommandInfo` one class scope further up is way
smoother.