Make spawn type enum protected to be used by controllers
Custom game modes can override `CanSpawn` and implement their own
spawn type picking.
This became harder because in #10846 a private enum was introduced.

https://github.com/ddnet/ddnet/pull/10846#issuecomment-3261701115

This change does not help ddnet. I would need that in my fork
ddnet-insta to avoid creating some kind of hack.
I would argue this fixes a regression introduced by #10846
and this change does not cause more maintenance for ddnet.

But it still falls under #7777 and can be rejected without the need for
any further justification.