Improve flag rendering performance, also without flags
Cache the current and previous `CNetObj_Flag` snap objects in the `CSnapState`, instead of iterating over all snap items in the current snapshot each time that items including flags are being rendered.

Also cache the previous `CNetObj_GameData` snap object in the `CSnapState` in addition to the current one, to avoid searching for this snap item multiple times. Thereby, the `m_GameDataSnapId` member is not necessary anymore.

On the map dm1 with 64 players at zoom 0, this decreases the average frametime from 1431 µs to 1180 µs. The CPU time spent in the `CItems::OnRender` function in this scenario is reduced from 9.14% to 1.02% of the total CPU time.