Improve chat rendering performance
Avoid iterating over and updating uninitialized chat lines every frame. Add member variable `bool m_Initialized` to the `CLine` class to make it explicit whether a chat line has been initialized. Also avoid uninitialized chat lines being considered when checking whether the current message is a repeat of the previous one.

On the map dm1 with 64 players at zoom 0 with the optimizations of #10864 as the baseline, this decreases the average frametime from 1164 µs to 1084 µs. The CPU time spent in the `CChat::OnPrepareLines` function in this scenario is reduced from 3.50% to 0.01% of the total CPU time.