Fix doxygen navigation structure
The `@mainpage` and `@page` commands already set the top-most page title for Doxygen. The `@page mastersrv mastersrv` was not correctly transformed while inside the `#`-headline with the older Doxygen version 1.9.8 that is used on the server.

The other `##`-headers are changed to `#` so their header level is directly below the page itself, otherwise they are not shown in the navigation menu.