Add important moderator and server alert messages
Alerts are supposed to be used only for important notifications relating to the server status and moderator actions that should not be missable by players. Important alerts are different from broadcasts in the following ways:

- Important alerts are larger and use have a title in red color, which is localized on the client-side.
- Important alerts are always shown to both the main and dummy player.
- Important alerts are shown for a longer duration (20 seconds).
- Important alerts cause broadcast and MOTD to be hidden.
- Important alerts send a notification (not supported on Windows yet) (notifications can be enabled/disabled with `cl_shownotifications`).
- Important alerts cannot be disabled with client settings.

The server command `moderator_alert <victim> <message>` shows an important alert with localized title `Moderator alert` to clients. Moderators would be given access to this command, to send unmissable warnings to specific players. As with most victim-based commands, using `all` is possible but strongly discouraged.

The server command `server_alert <message>` shows an important alert with localized title `Server alert` to all clients. Only admins would be given access to this command, to send important alerts like `The server will restart in 60 seconds.\nPlease save your run with /save`.

For older clients, that do not support the new netmessage, a broadcast message is sent as fallback. Using `moderator_alert` also causes a notice to be printed in the console for the moderator, that informs them that players using old clients might miss the alert.