Slightly improve player and freezebar performance
Use the current player info snapshot pointers that that are already stored in the snapshot state and also store pointers to the previous player info to avoid the repeated linear searches for the snapshot items when rendering players and freezebars.

Move the additional check `GameClient()->m_Snap.m_aCharacters[ClientId].m_Active` into the `IsPlayerInfoAvailable` functions to avoid duplicate code.

Profiling with 64 players with freezebars shows that the CPU time spent in the `CPlayers::OnRender` function is reduced from 5.99% to 3.87%. CPU time in the `CFreezeBars::OnRender` function is reduced from 2.93% to 1.97%. CPU time in the `CGameClient::OnNewSnapshot` function is increased from 0.69% to 0.72%. Frametimes are apparently not affected much by this change.