Cleanup 0.7 skin/part listbox selection handling
Determine the index of the selected skin/part in a separate loop and pass it to the `DoStart::DoStart` function to avoid the `static` variable for the selected index and fix the incorrect variable name `s_OldSelected` of a non-static variable.

Rename variable `s_paList` to `s_avpList`.

Remove clang-tidy exclusion for local variable names and constants with the prefix `s_`.