Fix performance regression due to very frequent skin updating
Skins were only supposed to be updated every 25 milliseconds or slower depending on the FPS. However, due to the incorrect type `std::chrono::microseconds` being used and the `IClient::RenderFrameTime` function actually returning the average frame time in seconds, the update duration was reduced to only 25 microseconds.

Adjusting the skin updating to a more realistic frequency reduces average frametimes from 895 us to 622 us (1117 to 1607 FPS). Profiling shows that the CPU time spent in the `CSkins::OnUpdate` function is reduced from 22.81% to 0.87%.