Overhaul controls settings: multiple binds, search, show custom, …
Support binding multiple key combinations for the same bind option in the controls settings. Show "add"-button for each bind option to bind another key. Show "remove"-button for each individual key binder to unbind that key combination for convenience. Right-clicking to unbind a key combination still works, but is less discoverable. Closes #10618.

Add search function for bind options. The search function considers the localized label of the bind option (e.g. `Move left`) as well as the command (e.g. `+left"`). The hotkey Ctrl+F activates the search. Pressing Tab/Enter navigates to the next match. Pressing Shift+Tab/Enter navigates to the previous match. The search input is positioned in a bottom bar outside of the scrollregion. Also move the "Reset to defaults" button to the bottom bar.

Show an additional bind group at the end of the list for "Custom" binds, which are binds for which no predefined bind option exists. The bind command is shown as the label for these bind options and they are listed in alphabetical order.

Make bind option groups collapsible by clicking on the title bar. All bind option groups except for the Custom binds are expanded by default.

Show ellipsis for labels of custom bind options (i.e. the commands) if they are too long. Show the full command in a tooltip instead, if the label was truncated. Also show the bind command for all predefined bind options to make the relevant commands more discoverable.

Various refactoring:
- Avoid all `static` variables for controls settings.
- Apply culling to some controls settings UI elements, i.e. only render and update them when they are inside of the viewport of the scrollregion, to improve rendering performance.
- Add constants for font sizes, button height and margins.
- First calculate the required height for collapsible settings (previously only joystick settings), then allocate required height, instead of clipping for one frame.
- Separate settings rendering into more functions (mouse, joystick, binds).
- Use `CBindSlot` instead of separate `int`s for key and modifier combination.
- Change key binder component to add separate button for unbinding. Allow activating key binder programatically. Also return whether the key binder was aborted with Escape. Use `std::optional` instead of separate `bool` to denote whether a key was read. Use `CBindSlot` instead of separate `IInput::CEvent` for key and `int` for modifier combination.