Also show favorite skins that were not found in list
Previously, you could only see favorite skins that could be loaded successfully. Now, all favorite skins are shown in the list, with question marks if they could not be loaded, e.g. if you favorite a skin that does not exists, or if you favorite a non-vanilla skin and enable vanilla-only.