Make clang-tidy ignored regex for classes and structs more specific
Previously, any struct name starting with `C` was accepted due to the regex matching only the prefix. Now, the entire name is matched to check for proper Camel Case after the prefix. The prefixes `I`, `S` and `GL_S` as well as underscores are also allowed to avoid many changes, but underscores and the prefixes `S` and `GL_S` should eventually be phased out when removing `struct`s entirely.