Fix client assertion crash on 0.7 servers with high ping
On 0.7 servers with higher ping, it takes a moment until the 0.7 skin info is available, which was previously causing the client to crash with an assertion because the 0.6 skin info was used as fallback but not initialized.

Now, the 0.6 skin is always initialized to `default` as fallback.

The skin info variables in the translation context are removed because they are unused.