Move sixup security tokens into the packet class
Makes the code cleaner because the tokens are now tied
to the packet object which they belong to.

Also gets rid of the need to pass two tokens separately
around as arguments.

Makes it clear that these tokens are used for sixup only
and are not to be confused with the ddnet security tokens.