Fix ninja old velocity not being clamped
When ninja runs out the tee's velocity is set to old velocity stored at the start of ninja, but this happens in HandleWeapon() which is after PreTick() so we don't clamp the velocity before MoveBox() happens, then MoveBox() runs a ~2 billion iteration loop.

Note: CCharacter::PreTick() does m_Core.Tick() which does CCharacterCore::TickDeferred(), which is when we clamp velocity magnitude to 6000,  (under default settings)

The MoveBox() in HandleNinja isn't affected by this because it just sets your speed to fixed value.

The reason the normal shotgun bug doesn't have this lag is because the velocity gets clamped to 6000 in PreTick(), before MoveBox() so you never actually move faster than 6000.

This doesn't change any physics in any other situation because there aren't any other situations where you get over 6000 speed so we can safely clamp velocity to 6000 anywhere at any time.

You could put this in a different place, but putting it here also fixes the stored velocity being negative when it overflows the int. (that's why the reproduction requires the ninja tee to look down, so the negative stored velocity moves you up)

See discord message for extra long explanation
https://discord.com/channels/252358080522747904/293493549758939136/1427568007290552382