Remove unnecessary special case when binding to left mouse button
Key binders were binding to the left mouse button when pressing the key, whereas all other keys are bound on key release. The special handling for the left mouse button is not necessary anymore, as key binder buttons are only activated after releasing the left mouse button, so they won't bind to the left mouse button for the same click that activated the binder.