Do not count player vote on client-side until server confirms it
The client was assuming that its vote was accepted when sending it. Now the voted-state is only updated when the client receives the `Sv_YourVote` message from the server. This might have been a left-over from the original implementation of #7181 or for compatibility with servers that don't send `Sv_YourVote`, but I prefer showing the correct state for servers that do send `Sv_YourVote` instead.

Closes #10653.