Print some of the variable values that failed an assert
This is not going throught ALL occurences of asserts.
I just picked a certain type of assert and unified the
error message while also enriching the assert information.

The goal of this commit is to establish the new style that
was intentionally picked in #11205 so future asserts can
copy that. There is no need to enforce this style in the CI
or anywhere else.

Another goal of this commit was to have the most useful error
message possible. So I did not touch existing asserts where
strictly applying the new style would be less descriptive.

Closed #11205