Fix invalid label for custom binds when rebinding to same command
The label of custom binds options was a pointer to the command string allocated by `CBinds`, which is freed when the command is unbound. Unused bind options were already removed automatically, but this did not handle the case that the same command is unbound and immediately bound again. Now, the label is not stored separately for custom bind options anymore, as the stored command is identical and already stored as an `std::string` to avoid this issue.