Fix incorrect seasonal theme being used due to thread-local time
In the `time_season` function, the time info is first fetched by calling `time_localtime_threadlocal`. If the following `if`-branches are not taken, such that `time_iseasterday` is called, then the time info that the `time_info` pointer points to will have been overridden from the `time_localtime_threadlocal` call in the `time_iseasterday` function.

To avoid this, the `time_localtime_threadlocal` function is changed to return a copy of the time info, so the returned info stays valid.

This was causing the wrong season to be detected in some cases. For example, on the 29th November the date that was used at the end of the `time_season` function was the 1st December, causing the winter theme to be used too early.