Assert that we never use negative values for ddrace teams
CC #11357

I see two places where `Team` is set by user input.
That is the chat command **/team** and the rcon command `set_team_ddr`.
Invalid user input should be dropped in the command.
And at any later point in time it should die with an assert because
the value should not be passed around when invalid.

Before this commit the rcon command `set_team_ddr` would silently fail
when the team is out of bounds.
This commit now shows a helpful error message to the rcon user instead.

Before this commit both the /join and /team command would try to search
for a free team when out of bounds. For the /join command that made
little sense and it also should have never happend.
So this out of bounds logic is now explicitly only moved to the /team
command.