Log stdout, stderr and Valgrind errors of failed integration tests
When an integration test failed, also log the Valgrind errors, instead of only logging them on test success.

Log the last 5 lines of stdout and stderr of all launched processes when an integration test failed or if Valgrind detected errors.

Add `--show-full-output` option to log full stdout and stderr to use in CI.