Mark `logfile` read only on the server side
The command has no effect when set from rcon anways.
Not even using `reload` helps to actually get it to log.

So being able to set it was misleading.

And we also probably do not want it to take effect when being set from
rcon. Because of security concerns (see #11376).