Add `-ffile-prefix-map` for more reproducible release builds
Replace absolute project path in macros like `__FILE__` with the name of the project.

This makes the build more reproducible and avoids showing the full project path in assertion error messages.

This is enabled in all non-debug build modes (`Release`, `RelWithDebInfo`, `MinSizeRel`) unless `DEV` is enabled.

The `DEV=TRUE` setting is removed from the Android build, so that the release build is more reproducible. It was previously already unnecessary, because only the `game-client` and `game-server` targets are built explicitly.