Move snapping of score into the controller
This commit falls under #7777 as the main motivation is
to allow forks of ddnet to change the score value
without having to create a git diff that has to maintained.

I have the same method in my fork already since a while and it allows multiple
gamemodes to set the value depending on their needs.
Because #7777 was rejected it is not upstreamed to ddnet.

I actually tried to add this very method before in #7598
because it is such a key requirement for custom gametypes.
And a common painpoint for modders like me that try to stay close to
ddnet upstream.

My interested in proposing this again was caused by the length of this
line:

https://github.com/ddnet/ddnet/blob/77c17304eb0a160728b17e932aa215a2b524b66e/src/game/server/player.cpp#L385

```C++
pPlayerInfo->m_Score = m_Score.has_value() && (!g_Config.m_SvHideScore || SnappingClient == m_ClientId) ? GameServer()->Score()->PlayerData(m_ClientId)->m_BestTime * 1000 : protocol7::FinishTime::NOT_FINISHED;
```

And the score being set in 3 different places in the player snap method.
So this time I would argue that ddnet it self directly benfits from this
cleanup.