Minor improvement of client input validation on server-side
Reject inputs if the size is not aligned with `int32_t`s as expected.

Reject inputs if the last acked snapshot is out of range. The value `-1` is used when the client first connects and no snapshot has been acked. Other negative values are not expected. Values above the maximum tick are also not expected.

Avoid trying to retrieve a snapshot with negative tick from storage when the last acked snapshot is still unset, as the storage is either empty (with regular clients) or this causes an unnecessary linear search for an entry that the snapshot storage cannot contain (with modified clients).

Use `sizeof(int32_t)` instead of `4`.