Improve client input validation on server-side
Reject inputs if the vital-flag is set. Inputs are always non-vital.

Reject inputs for clients that are not ingame. Previously, the direct input was not passed to the game server for clients that are not ingame, but there is no need to handle the input at all, as clients are not expected to send any input until they are ingame.

Reject inputs if the size is not aligned with `int32_t`s as expected, or if the size is smaller than `CNetObj_PlayerInput` has been for the last 15 years.

Reject inputs if the last acked snapshot is out of range. The value `-1` is used when the client first connects and no snapshot has been acked. Other negative values are not expected. Values above the current tick are also not expected, as the client cannot acknowledged a snapshot from a future tick.

Avoid sending preinputs for inputs whose intended tick is more than 4 seconds into the future.

Avoid sending preinputs to clients that are not in full snapshot mode, to allow them to recover and reduce the amount of optional messages sent to them.

Avoid trying to retrieve a snapshot with negative tick from the snapshot storage when the last acked snapshot is still unset, as the snapshot storage is either empty (with regular clients) or this causes an unnecessary linear search for an entry that the snapshot storage cannot contain (with modified clients).

Use `sizeof(int32_t)` instead of `4`.