Wait for both clients to connect before starting race in smoke test
Perform minimal change of coverage map to ensure that the first client does not start the race until the second client is connected, by blocking the path of client 1 with a switch that is deactivated when the client 2 spawns.

Otherwise, it's possible for the second client to not be connected in time when the first client waits at the door between checkpoints 18 and 19.

This should fix the ranks being mismatched when using Valgrind.