Only save settings for first client in smoke test
Quitting both clients at the same time can cause them to rename the temporary config file to the real filename at same time on Windows, which will fail for one client. Failing to save the config shows a popup error message, that prevents the client from quitting and causes the smoke test to fail waiting for the client to exit.

Alternative to #11506.