Add Emscripten libraries, build Emscripten client in CI
Add precompiled libraries for Emscripten to the ddnet-libs repository. The libraries are built according to the README using the scripts updated in #11497. Emscripten version 4.0.22 was used to build them. The library build is reproducible on Linux with a fixed version of the Emscripten SDK.

Add a manually dispatchable GitHub workflow to build the Emscripten libraries on an Ubuntu runner. This can be used to verify that the library build is reproducible.

Add workflow to build the Emscripten client in the CI. First build the headless Emscripten client in debug mode and run it in headless Firefox. Then build the Emscripten client with graphics in release mode and upload it as an artifact. The release client build is also reproducible without any additional efforts when using fixed versions of Rust (1.89.0 in the CI) and CMake (3.22.1 in the CI).

Support stdout/stderr of Emscripten client being captured by the `emrun` tool in the CI by adding the linker flag `--emrun` when building in debug mode.